+++ docs/locales/en-US/index.ts
@@ -56,7 +56,8 @@ export default {
     'tab.summary': 'Summary',
     'tab.fixed': 'Fixed Column',
     'tab.auto-height': 'Auto Height',
-    'tab.affix-header': 'Affix Header'
+    'tab.affix-header': 'Affix Header',
+    'tab.affix-horizontal-scrollbar': 'Affix Horizontal Scrollbar'
   },
   palette: {
     title: 'Get the product color',