+++ src/MultiCascader/DropdownMenu.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import PropTypes from 'prop-types';
 import _ from 'lodash';
 import classNames from 'classnames';
-import { shallowEqual } from 'rsuite-utils/lib/utils';
+import shallowEqual from '../utils/shallowEqual';
 
 import { getUnhandledProps, prefix } from '../utils';
 import { DropdownMenuCheckItem } from '../Picker';