+++ src/Tooltip/Tooltip.tsx
@@ -8,8 +8,6 @@ import { overlayProps } from '../Whisper/Whisper';
 
 class Tooltip extends React.Component<TooltipProps> {
   static propTypes = {
-    positionLeft: PropTypes.number,
-    positionTop: PropTypes.number,
     visible: PropTypes.bool,
     classPrefix: PropTypes.string,
     className: PropTypes.string,
@@ -19,26 +17,29 @@ class Tooltip extends React.Component<TooltipProps> {
   render() {
     const {
       className,
-      positionLeft,
-      positionTop,
       classPrefix,
       children,
       style,
       visible,
+      htmlElementRef,
       ...rest
     } = this.props;
 
     const addPrefix = prefix(classPrefix);
     const classes = classNames(classPrefix, className);
     const styles = {
-      left: positionLeft,
-      top: positionTop,
       opacity: visible ? 1 : undefined,
       ...style
     };
 
     return (
-      <div {..._.omit(rest, overlayProps)} role="tooltip" className={classes} style={styles}>
+      <div
+        {..._.omit(rest, overlayProps)}
+        role="tooltip"
+        className={classes}
+        style={styles}
+        ref={htmlElementRef}
+      >
         <div className={addPrefix('arrow')} />
         <div className={addPrefix('inner')}>{children}</div>
       </div>