+++ src/TreePicker/TreePicker.d.ts
@@ -2,7 +2,21 @@ import * as React from 'react';
 
 import { FormControlPickerProps } from '../@types/common';
 import { TreeBaseProps } from '../Tree/TreeBase';
+import { TREE_NODE_DROP_POSITION } from '../constants';
 
+export interface DropData {
+  /** drag node data */
+  dragNode: any;
+
+  /** dropNode data */
+  dropNode: any;
+
+  /** node drop postion */
+  dropNodePosition: TREE_NODE_DROP_POSITION;
+
+  /** Update Data when drop node */
+  createUpdateDataFunction: (data: any[]) => any[];
+}
 export interface TreePickerProps extends TreeBaseProps, FormControlPickerProps {
   /** The height of Dropdown */
   height?: number;
@@ -18,6 +32,9 @@ export interface TreePickerProps extends TreeBaseProps, FormControlPickerProps {
 
   /** Whether using virtualized list */
   virtualized?: boolean;
+
+  /** Whether the node can  be dragged */
+  draggable?: boolean;
 
   /** Set the option value for the expand node */
   defaultExpandItemValues?: any[];
@@ -37,6 +54,26 @@ export interface TreePickerProps extends TreeBaseProps, FormControlPickerProps {
 
   /** Called when scrolling */
   onScroll?: (event: React.SyntheticEvent<HTMLElement>) => void;
+
+  /** Called when node drag start */
+  onDragStart?: (nodeData: any, e: React.DragEvent) => void;
+
+  /** Called when node drag enter */
+  onDragEnter?: (nodeData: any, e: React.DragEvent) => void;
+
+  /** Called when node drag over */
+  onDragOver?: (nodeData: any, e: React.DragEvent) => void;
+
+  /** Called when node drag leave */
+  onDragLeave?: (nodeData: any, e: React.DragEvent) => void;
+
+  /** Called when node drag end */
+  onDragEnd?: (nodeData: any, e: React.DragEvent) => void;
+
+  /** Called when node drop */
+  onDrop?: (dropData: DropData, e: React.DragEvent) => void;
+
+  renderDragNode?: (dragNode: any) => React.ReactNode;
 }
 
 declare const TreePicker: React.ComponentType<TreePickerProps>;