+++ CHANGELOG.md
@@ -1,3 +1,91 @@
+# 4.4.1
+
+> April 29, 2020
+
+- **Bugfix**: Fix the clear button of picker not clickable in IE11. ([#1017])
+- **Bugfix**: Fix toggle clean button styles issue. ([#1016])
+- **Bugfix**: Fix missing typescript type definitions. ([#1008])
+- **Bugfix**: Fix the expanded state of the menu. ([#1007])
+
+---
+
+- **Bugfix**: 修复了 IE11 的一个兼容性问题，Picker 的清除按钮不可点击 ([#1017])
+- **Bugfix**: 修复了 Toggle 按钮样式问题。 ([#1016])
+- **Bugfix**: 修复了遗漏的 Typescript 类型定义。 ([#1008])
+- **Bugfix**: 修复了 Dropdown 中的菜单状态错误的问题。([#1007])
+
+[#1017]: https://github.com/rsuite/rsuite/pull/1017
+[#1016]: https://github.com/rsuite/rsuite/pull/1016
+[#1008]: https://github.com/rsuite/rsuite/pull/1008
+[#1007]: https://github.com/rsuite/rsuite/pull/1007
+
+# 4.4.0
+
+> April 24, 2020
+
+- **Feature**: Support `draggable` on `<Tree>` ([#987])
+- **Feature**: Added `<Rete>`. ([#981])
+- **Feature**: Added support for Italian ([#971])
+- **Feature**: Added support for custom overlay. ([#969])
+- **Feature**: Support `endless` on `<Timeline>` ([#982])
+- **Feature**: Support `vertical` on `<Progress.Line>`. ([#938])
+- **Feature**: Support `showMeridian` on `<DatePicker>`. ([#932])
+- **Feature**: Support `searchBy` on all picker. ([#928],[#965])
+- **Feature**: Support `tagProps` on `<TagPicker>`. ([#927])
+- **Feature**: Support `container` on `<Affix>`. ([#912])
+- **Feature**: Support `onDataUpdated` on `<Table>`. ([rsuite-table#140])
+- **Feature**: Support `affixHorizontalScrollbar` on `<Table>`. ([rsuite-table#137])
+- **Breaking**: Fix misspelling of Uploader props `dragable`, modify to`draggable`.
+- **Improve**: Picker support size props at default. ([#917])
+- **Improve** Improved animation of `<Placeholder>` active state.
+- **Improve**: Improved `wordWrap` property for compatibility with `rowHeight`. ([rsuite-table#138])
+- **Bugfix**: Fix Types.PlacementAuto definition. ([#962])
+- **Chore**: Migrating `rsuite-utils` to `rsuite`. ([#963])
+- **Example**: Add with-preact example. ([#977])
+
+---
+
+- **Feature**: `<Tree>` 组件新增 `draggable`属性， 支持拖拽。 ([#987])
+- **Feature**: 新增 `<Rete>` 组件。 ([#981])
+- **Feature**: 新增对意大利语言支持。 ([#971])
+- **Feature**: 支持自定义浮层。 ([#969])
+- **Feature**: `<Timeline>` 组件新增 `endless` 属性。 ([#982])
+- **Feature**: `<Progress.Line>` 组件新增 `vertical` 属性，垂直显示进度条。 ([#938])
+- **Feature**: `<DatePicker>` 组件新增 `showMeridian` 属性，支持时间 12 小时格式显示。 ([#932])
+- **Feature**: `<TagPicker>` 组件支持 `tagProps`，设置 Tag 属性。 ([#927])
+- **Feature**: `<Affix>` 支持 `container` 属性，把元素只在容器可见范围内才固定。 ([#912])
+- **Feature**: `<Table>` 支持 `onDataUpdated` 属性，在表格数据加载完成以后的回调函数。 ([rsuite-table#140])
+- **Feature**: `<Table>` 支持 `affixHorizontalScrollbar` 属性，让表格底部横向滚动条固定在可视范围内。 ([rsuite-table#137])
+- **Feature**: 所有带搜索功能的 Picker 组件 `searchBy`属性，可以自定义搜索规则。 ([#928],[#965])
+- **Breaking**: 修复 Uploader 属性 `dragable` 拼写错误，修改为 `draggable`。 ([#984])
+- **Improve**: 改进所有的 Picker 组件，默认支持 `size` 属性。([#917])
+- **Improve** 改进 `<Placeholder>` 的动画效果。
+- **Bugfix**: 修复 `<Table>` 的 `wordWrap` 属性与 `rowHeight`不兼容的问题。 ([rsuite-table#138])
+- **Bugfix**: 修复 Typescript 一些定义错误。([#962])
+- **Example**: 新增示例项目 with-preact。 ([#977])
+- **Chore**: 迁移 `rsuite-utils` 库到 `rsuite`，方便维护。 ([#963])
+
+[#987]: https://github.com/rsuite/rsuite/pull/987
+[#984]: https://github.com/rsuite/rsuite/pull/984
+[#987]: https://github.com/rsuite/rsuite/pull/982
+[#981]: https://github.com/rsuite/rsuite/pull/981
+[#977]: https://github.com/rsuite/rsuite/pull/977
+[#971]: https://github.com/rsuite/rsuite/pull/971
+[#969]: https://github.com/rsuite/rsuite/pull/969
+[#965]: https://github.com/rsuite/rsuite/pull/965
+[#963]: https://github.com/rsuite/rsuite/pull/963
+[#962]: https://github.com/rsuite/rsuite/pull/962
+[#938]: https://github.com/rsuite/rsuite/pull/938
+[#933]: https://github.com/rsuite/rsuite/pull/933
+[#932]: https://github.com/rsuite/rsuite/pull/932
+[#928]: https://github.com/rsuite/rsuite/pull/928
+[#927]: https://github.com/rsuite/rsuite/pull/927
+[#917]: https://github.com/rsuite/rsuite/pull/917
+[#912]: https://github.com/rsuite/rsuite/pull/912
+[rsuite-table#140]: https://github.com/rsuite/rsuite-table/pull/140
+[rsuite-table#138]: https://github.com/rsuite/rsuite-table/pull/138
+[rsuite-table#137]: https://github.com/rsuite/rsuite-table/pull/137
+
 # 4.3.4
 
 > April 10, 2020
@@ -197,7 +285,7 @@
 - **Feature**: Added support for `parentSelectable` on `<Cascader>`. ([#802])
 - **Feature**: Added support for Russian. ([#799])
 - **Feature**: Added option `image` to the `graph` property of `<Placeholder>` ([#755])
-- **Feature**: Added support for `dragable` on `<Uploader>`. ([#752])
+- **Feature**: Added support for `draggable` on `<Uploader>`. ([#752])
 - **Improve**: Enhance `<Tooltip>` and `<Popover>` to support HTML attributes ([#806])
 - **Bugfix**: Fixed the height of the buttons in the `<InputGroup>`.([#807])
 - **Bugfix**: Fixed `Drawer` gap bug when in RTL mode.([#803])
@@ -211,7 +299,7 @@
 - **Feature**: `<Cascader>` 组件支持 `parentSelectable` 属性，让父节点可选择。([#802])
 - **Feature**: 添加对俄语支持。([#799])
 - **Feature**: `<Placeholder>`的 `graph` 属性值添加了 `image` 选项。 ([#755])
-- **Feature**: `<Uploader>` 组件支持 `dragable` 属性，可以拖拽上传文件。([#752])
+- **Feature**: `<Uploader>` 组件支持 `draggable` 属性，可以拖拽上传文件。([#752])
 - **Improve**: `<Tooltip>` 和 `<Popover>` 支持 HTML 元素默认属性。 ([#806])
 - **Bugfix**: 修复了 `<InputGroup>` 内部按钮高度的问题。([#807])
 - **Bugfix**: 修复了 `<Drawer>` 在 RTL 模式下，左侧存在间隙的样式问题。([#803])